0x03 task13 len of list returns an integer

+++ 0x03-python-data_structures/13-is_palindrome.c
@@ -6,10 +6,10 @@
  * Return: The number of elements (Nodes)
  */
 
-size_t list_len(const listint_t *h)
+int list_len(const listint_t *h)
 {
 	const listint_t *current = NULL;
-	size_t counter = 0;
+	int counter = 0;
 
 	current = h;
 	while (current != NULL)
